Add service descriptions.
Add a long and a short description for each service.  The short one is
used in the service list (and show in the admin overview).  The long
one is showed on the service page and contains markdown (so it can
e.g. link to the actual service).

Also add number of created accounts to admin overview.