Add a staff_or_superuser_required decorator
This decorator can be used to restrict access to certain views.

One reason for adding this is to reduce code duplication and ensure that
we do the same restriction on all pages where we want this type of
restriction.

It is slightly different from the existing blocks in `finger.views` in
that it returns a `HttpResponseForbidden` instead of a
`HttpResponseNotAllowed`. The reason for this is that forbidden returns
HTTP 403, while not allowed is actually HTTP 405 (Method Not Allowed).

Method Not Allowed should be used to say that the request method used in
the request is not allowed for this particular URL. That is not what we
want to signal here. We want to say that the user session trying to
reach this URL does not have permission.

The reason I use HTTP 403 instead of HTTP 401 (Unauthorized) is because
401 signals that authentication is possible to achieve access.

The decorator itself uses `login_required`, which will return
`HttpResponseUnauthorized` if there is no user so we don't need to test
for that in our code.