Do no delete an existing admin user
The old code always deleted the admin user and re-created a new one.
This had the effect that the user was logged out every single time
(because the user id was removed). This also spammed the database with
user id:s.

The password will be updated if an admin user already exists.