A finger.json entry now refers to fingerweb
There was to little data to safely identify and map a finger.json entry
to a fingerweb Member object. This commit adds an identifier
(fingerweb_identifier inside finger.json) entry for that.

Most entries do not need the new identifier, I can use the username,
email or name to uniquely pinpoint a user.