Make stringified User/Member objects readable
When looking at the user admin page, the member select dropdown looks
something like this:

    Member object (1)
    Member object (2)
    Member object (3)
    Member object (4)
    Member object (5)

This is a bit hard to read, and while it is possible to check exactly
which member a specific user should be connected to, it is tedious and
requires the admin to memorize database primary keys.

Adding this `__str__` methods to the `Member` model will make the list
contain the members' full name (and db primary key) instead.