Send any template and various tweaks
Update email confirmation page to include a template selector. Also
rename "Send Welcome E-mail" to "Send E-mail" in the dropdown to reflect
the now more generic use.

Also send email to user@stacken.kth.se, is the member has a user. Also
include the full name in the email for nicer rendering.

Set the sender to the logged in user, and set the name to the classic
"Datorföreningen Stacken via ..." message that send.py used.